预约转诊、服务权限逻辑更改，实时判断 预约服务、转诊身份证号码校验 banner无网络处理

+++ frame/src/main/java/com/yht/frame/data/bean/SelectCheckTypeBean.java
@@ -11,17 +11,17 @@ import java.io.Serializable;
  */
 public class SelectCheckTypeBean extends DataSupport implements Serializable {
     private static final long serialVersionUID = 2307028291664717958L;
-    private int price;
+    private long price;
     private String projectCode;
     private String projectName;
     private String hospitalName;
     private String hospitalCode;
 
-    public int getPrice() {
+    public long getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(long price) {
         this.price = price;
     }
